+++ packages/client/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @nifrajs/client
 
+## 1.1.0
+
 ## 1.0.0
 
 ### Minor Changes